HLSv3 EXTINF changes.
Fail on parsing a playlist that uses floats for EXTINF times with a
defined version of V2. Could see adding a lax mode for bad producers, and fix on output, but
for now be strict.

Supresses io.EOF errors on manifest/hls for Parser. Eventually a user
would expect for an EOF. If there is a case where EOF was encountered
early, it should be caught when we Encode the output.